+++ src/modules/add.js
@@ -1,16 +1,14 @@
-const myArray = [];
-
-export default class User {
-  constructor(name, score) {
-    this.name = name;
-    this.score = score;
-  }
-}
-
-const addToArray = (name, score) => {
-  const userScore = new User(name, score);
-  myArray.push(userScore);
-  localStorage.setItem('userScores', JSON.stringify(myArray));
+const url = 'https://us-central1-js-capstone-backend.cloudfunctions.net/api/games/b1GI7GiQGpLXMcZYR0Tj/scores';
+const Post = async (user, score) => {
+  const res = await fetch(url, {
+    method: 'POST',
+    headers: {
+      'Content-Type': 'application/json',
+    },
+    body: JSON.stringify({ user, score }),
+  });
+  const data = res.json();
+  return data;
 };
 
-export { addToArray, myArray };
+export { Post, url };
