--- src/modules/display.js
@@ -1,8 +1,0 @@
-const tablerow = document.querySelector('userData');
-const displayInfo = (name, score) => {
-  const row = document.createElement('tr');
-  row.textContent = `${name}: ${score}`;
-  tablerow.appendChild(row);
-};
-
-export default displayInfo;
